pyproject.toml
[project]
name = "smart-home-mcp"
version = "0.2.0"
description = "MCP Server for controlling smart home devices from multiple brands"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
midea = ["midea-beautiful-air>=0.10.5"]
all = ["smart-home-mcp[midea]"]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/smart_home_mcp"]
